insertion in sql

+++ 01_basics/migrations/001_students_table.up.sql
@@ -1,2 +1,2 @@
 ALTER TABLE students
-ADD COLUMN phone INTEGER;
+ADD COLUMN phone BIGINT;
